dom/media/webaudio/test/test_oscillatorNodePassThrough.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Oscillator with passthrough</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var gTest = {
  length: 2048,
  numberOfChannels: 1,
  createGraph: function(context) {
    var buffer = context.createBuffer(1, 2048, context.sampleRate);
    for (var i = 0; i < 2048; ++i) {
      buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
    }

    var source = context.createOscillator();

    var srcWrapped = SpecialPowers.wrap(source);
    ok("passThrough" in srcWrapped, "OscillatorNode should support the passThrough API");
    srcWrapped.passThrough = true;

    source.start(0);
    return source;
  },
  createExpectedBuffers: function(context) {
    var expectedBuffer = context.createBuffer(1, 2048, context.sampleRate);

    return [expectedBuffer];
  },
};

runTest();

</script>
</pre>
</body>
</html>
